Fix chart installation against Tiller 2.4.1
Chart installation against Tiller 2.4.1 was broken because the error
message returned contained the quoted name of the release being
installed. The previous code checked a very specific error message to
determine if it should install or update. This fix should be backwards
compatable, though I'm sure there's a better way to do it that I don't
see right now.